# RecruitmentAgency/RecruitmentAgency.WASM/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RecruitmentAgency.API",
    "version": "1.0"
  },
  "paths": {
    "/api/ApplicantApplications": {
      "get": {
        "tags": [
          "ApplicantApplications"
        ],
        "summary": "Получает все заявки соискателей.",
        "responses": {
          "200": {
            "description": "Заявки успешно получены.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantApplicantDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantApplicantDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantApplicantDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApplicantApplications"
        ],
        "summary": "Создает новую заявку соискателя.",
        "requestBody": {
          "description": "Данные для создания новой заявки.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantApplicationCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantApplicationCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantApplicationCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Заявка успешно создана.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              }
            }
          },
          "400": {
            "description": "Данные заявки недействительны."
          }
        }
      },
      "put": {
        "tags": [
          "ApplicantApplications"
        ],
        "summary": "Обновляет информацию о существующей заявке.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Идентификатор заявки соискателя.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Обновлённая информация о заявке.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantApplicationCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantApplicationCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantApplicationCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Заявка успешно обновлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              }
            }
          },
          "404": {
            "description": "Заявка с указанным идентификатором не найдена."
          }
        }
      }
    },
    "/api/ApplicantApplications/{id}": {
      "get": {
        "tags": [
          "ApplicantApplications"
        ],
        "summary": "Получает заявку соискателя по указанному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор заявки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заявка успешно найдена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicantDTO"
                }
              }
            }
          },
          "404": {
            "description": "Заявка с указанным идентификатором не найдена."
          }
        }
      },
      "delete": {
        "tags": [
          "ApplicantApplications"
        ],
        "summary": "Удаляет заявку соискателя по указанному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор заявки, которую необходимо удалить.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Заявка успешно удалена."
          },
          "404": {
            "description": "Заявка с указанным идентификатором не найдена."
          }
        }
      }
    },
    "/api/Applicants": {
      "get": {
        "tags": [
          "Applicants"
        ],
        "summary": "Получает список всех соискателей.",
        "responses": {
          "200": {
            "description": "Соискатели успешно получены.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Applicants"
        ],
        "summary": "Добавляет нового соискателя.",
        "requestBody": {
          "description": "Данные для создания нового соискателя.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              }
            }
          },
          "201": {
            "description": "Соискатель успешно создан."
          },
          "404": {
            "description": "Данные соискателя недействительны."
          }
        }
      }
    },
    "/api/Applicants/{id}": {
      "get": {
        "tags": [
          "Applicants"
        ],
        "summary": "Получает информацию о соискателе по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор соискателя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Соискатель успешно найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              }
            }
          },
          "404": {
            "description": "Соискатель с указанным идентификатором не найден."
          }
        }
      },
      "put": {
        "tags": [
          "Applicants"
        ],
        "summary": "Обновляет информацию о существующем соискателе.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор соискателя, которого нужно обновить.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Обновлённая информация о соискателе.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Соискатель успешно обновлён.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantDTO"
                }
              }
            }
          },
          "404": {
            "description": "Соискатель с указанным идентификатором не найден."
          }
        }
      },
      "delete": {
        "tags": [
          "Applicants"
        ],
        "summary": "Удаляет соискателя по указанному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор соискателя, которого нужно удалить.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Соискатель успешно удалён."
          },
          "404": {
            "description": "Соискатель с указанным идентификатором не найден."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicantApplicantDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор заявки соискателя",
            "format": "int32"
          },
          "submissionDate": {
            "type": "string",
            "description": "Дата подачи",
            "format": "date-time"
          },
          "applicantId": {
            "type": "integer",
            "description": "Соискатель",
            "format": "int32"
          },
          "positionId": {
            "type": "integer",
            "description": "Должность",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO заявки соискателя"
      },
      "ApplicantApplicationCreateDTO": {
        "type": "object",
        "properties": {
          "submissionDate": {
            "type": "string",
            "description": "Дата подачи",
            "format": "date-time"
          },
          "applicantId": {
            "type": "integer",
            "description": "Соискатель",
            "format": "int32"
          },
          "positionId": {
            "type": "integer",
            "description": "Должность",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания заявки соискателя"
      },
      "ApplicantCreateDTO": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "ФИО соискателя",
            "nullable": true
          },
          "contactInformation": {
            "type": "string",
            "description": "Телефон",
            "nullable": true
          },
          "experience": {
            "type": "number",
            "description": "Опыт работы",
            "format": "double"
          },
          "education": {
            "type": "string",
            "description": "Образование",
            "nullable": true
          },
          "salaries": {
            "type": "integer",
            "description": "Ожидаемый уровень зарплаты",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания организации"
      },
      "ApplicantDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор соискателя",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "description": "ФИО соискателя",
            "nullable": true
          },
          "contactInformation": {
            "type": "string",
            "description": "Телефон",
            "nullable": true
          },
          "experience": {
            "type": "number",
            "description": "Опыт работы",
            "format": "double"
          },
          "education": {
            "type": "string",
            "description": "Образование",
            "nullable": true
          },
          "salaries": {
            "type": "integer",
            "description": "Ожидаемый уровень зарплаты",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO Соискателя"
      },
...
    }
  }
}